+++ CHANGELOG.md
@@ -6,6 +6,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.14.0] - 2024-01-10
+
+### Added
+
+- `:RustLsp renderDiagnostic` command:
+  Render diagnostics as displayed during `cargo build`.
+- Add `Open in split` action to `explainError` and `renderDiagnostic`
+  float preview windows.
+
 ## [3.13.1] - 2024-01-10
 
 ### Fixed